Edit migrations

+++ db/migrate/20221003130100_create_food.rb
@@ -1,11 +1,10 @@
 class CreateFood < ActiveRecord::Migration[7.0]
   def change
     create_table :foods do |t|
-
       t.string :name
-      t.integer :measurement_unit 
-      t.double :price 
-      t.double :quantity
+      t.integer :measurement_unit
+      t.decimal :price
+      t.decimal :quantity
       t.timestamps
     end
   end
